+++ examples/TwoRotaryEncoders/TwoRotaryEncoders.ino
@@ -41,12 +41,12 @@ RotaryEncoder rotaryEncoder1( RE1_DI_ENCODER_A, RE1_DI_ENCODER_B, RE1_DI_ENCODER
 RotaryEncoder rotaryEncoder2( RE2_DI_ENCODER_A, RE2_DI_ENCODER_B, RE2_DI_ENCODER_SW, RE2_DO_ENCODER_VCC );
 
 
-void printKnob1Value( int value )
+void printKnob1Value( long value )
 {
 	Serial.printf( "RE1 value: %i\n", value );
 }
 
-void printKnob2Value( int value )
+void printKnob2Value( long value )
 {
 	Serial.printf( "RE2 value: %i\n", value );
 }
